Move models around again
 - Move machine related models to cura.Machines.Models
 - Move printer device related models to cura.PrinterOutput.Models
 - Other UI/GUI related modules in cura.UI
