+++ cura/CuraApplication.py
@@ -48,7 +48,7 @@ from cura.Arranging.Arrange import Arrange
 from cura.Arranging.ArrangeObjectsJob import ArrangeObjectsJob
 from cura.Arranging.ArrangeObjectsAllBuildPlatesJob import ArrangeObjectsAllBuildPlatesJob
 from cura.Arranging.ShapeArray import ShapeArray
-from cura.UI.GlobalStacksModel import GlobalStacksModel
+from cura.Machines.Models.GlobalStacksModel import GlobalStacksModel
 from cura.Scene.ConvexHullDecorator import ConvexHullDecorator
 from cura.Operations.SetParentOperation import SetParentOperation
 from cura.Scene.SliceableObjectDecorator import SliceableObjectDecorator
@@ -66,19 +66,18 @@ from UM.Settings.SettingFunction import SettingFunction
 from cura.Settings.CuraContainerRegistry import CuraContainerRegistry
 from cura.Settings.MachineNameValidator import MachineNameValidator
 
-from cura.UI.MachineModels.BuildPlateModel import BuildPlateModel
-from cura.UI.MachineModels.NozzleModel import NozzleModel
-from cura.UI.MachineModels.QualityProfilesDropDownMenuModel import QualityProfilesDropDownMenuModel
-from cura.UI.MachineModels.CustomQualityProfilesDropDownMenuModel import CustomQualityProfilesDropDownMenuModel
-from cura.UI.MachineModels.MultiBuildPlateModel import MultiBuildPlateModel
-from cura.UI.MachineModels.FavoriteMaterialsModel import FavoriteMaterialsModel
-from cura.UI.MachineModels.GenericMaterialsModel import GenericMaterialsModel
-from cura.UI.MachineModels.MaterialBrandsModel import MaterialBrandsModel
-from cura.UI.MachineModels.QualityManagementModel import QualityManagementModel
-from cura.UI.MachineModels.QualitySettingsModel import QualitySettingsModel
-from cura.UI.MachineModels.MachineManagementModel import MachineManagementModel
-
-from cura.UI.MachineModels.SettingVisibilityPresetsModel import SettingVisibilityPresetsModel
+from cura.Machines.Models.BuildPlateModel import BuildPlateModel
+from cura.Machines.Models.NozzleModel import NozzleModel
+from cura.Machines.Models.QualityProfilesDropDownMenuModel import QualityProfilesDropDownMenuModel
+from cura.Machines.Models.CustomQualityProfilesDropDownMenuModel import CustomQualityProfilesDropDownMenuModel
+from cura.Machines.Models.MultiBuildPlateModel import MultiBuildPlateModel
+from cura.Machines.Models.FavoriteMaterialsModel import FavoriteMaterialsModel
+from cura.Machines.Models.GenericMaterialsModel import GenericMaterialsModel
+from cura.Machines.Models.MaterialBrandsModel import MaterialBrandsModel
+from cura.Machines.Models.QualityManagementModel import QualityManagementModel
+from cura.Machines.Models.QualitySettingsModel import QualitySettingsModel
+from cura.Machines.Models.MachineManagementModel import MachineManagementModel
+from cura.Machines.Models.SettingVisibilityPresetsModel import SettingVisibilityPresetsModel
 
 from cura.Machines.MachineErrorChecker import MachineErrorChecker
 
@@ -101,8 +100,8 @@ from cura.TaskManagement.OnExitCallbackManager import OnExitCallbackManager
 
 from cura.Settings.MachineManager import MachineManager
 from cura.Settings.ExtruderManager import ExtruderManager
-from cura.UI.UserChangesModel import UserChangesModel
-from cura.UI.ExtrudersModel import ExtrudersModel
+from cura.Machines.Models.UserChangesModel import UserChangesModel
+from cura.Machines.Models.ExtrudersModel import ExtrudersModel
 from cura.Settings.MaterialSettingsVisibilityHandler import MaterialSettingsVisibilityHandler
 from cura.Settings.ContainerManager import ContainerManager
 from cura.Settings.SidebarCustomMenuItemsModel import SidebarCustomMenuItemsModel
@@ -111,7 +110,7 @@ from cura.Settings.CuraFormulaFunctions import CuraFormulaFunctions
 
 from cura.UI.ObjectsModel import ObjectsModel
 
-from cura.UI.MachineModels.DiscoveredPrintersModel import DiscoveredPrintersModel
+from cura.Machines.Models.DiscoveredPrintersModel import DiscoveredPrintersModel
 
 from cura.PrinterOutput.PrinterOutputDevice import PrinterOutputDevice
 from cura.PrinterOutput.NetworkMJPGImage import NetworkMJPGImage
