+++ cura/Machines/Models/GenericMaterialsModel.py
@@ -1,10 +1,10 @@
 # Copyright (c) 2018 Ultimaker B.V.
 # Cura is released under the terms of the LGPLv3 or higher.
 
-from cura.UI.MachineModels.BaseMaterialsModel import BaseMaterialsModel
+from cura.Machines.Models.BaseMaterialsModel import BaseMaterialsModel
 
-##  Model that shows the list of favorite materials.
-class FavoriteMaterialsModel(BaseMaterialsModel):
+class GenericMaterialsModel(BaseMaterialsModel):
+
     def __init__(self, parent = None):
         super().__init__(parent)
         self._update()
@@ -25,14 +25,14 @@ class FavoriteMaterialsModel(BaseMaterialsModel):
             if bool(metadata.get("removed", False)):
                 continue
 
-            # Only add results for favorite materials
-            if root_material_id not in self._favorite_ids:
+            # Only add results for generic materials
+            if metadata["brand"].lower() != "generic":
                 continue
 
             item = self._createMaterialItem(root_material_id, container_node)
             item_list.append(item)
 
         # Sort the item list alphabetically by name
-        item_list = sorted(item_list, key = lambda d: d["brand"].upper())
+        item_list = sorted(item_list, key = lambda d: d["name"].upper())
 
         self.setItems(item_list)
