+++ cura/Machines/Models/GlobalStacksModel.py
