+++ cura/PrinterOutput/GenericOutputController.py
@@ -5,13 +5,13 @@ from typing import TYPE_CHECKING, Set, Union, Optional
 
 from PyQt5.QtCore import QTimer
 
-from cura.PrinterOutput.PrinterOutputController import PrinterOutputController
+from .PrinterOutputController import PrinterOutputController
 
 if TYPE_CHECKING:
-    from cura.UI.PrintJobOutputModel import PrintJobOutputModel
-    from cura.UI.PrinterOutputModel import PrinterOutputModel
-    from cura.PrinterOutput.PrinterOutputDevice import PrinterOutputDevice
-    from cura.UI.ExtruderOutputModel import ExtruderOutputModel
+    from .Models.PrintJobOutputModel import PrintJobOutputModel
+    from .Models.PrinterOutputModel import PrinterOutputModel
+    from .PrinterOutputDevice import PrinterOutputDevice
+    from .Models.ExtruderOutputModel import ExtruderOutputModel
 
 
 class GenericOutputController(PrinterOutputController):
