+++ plugins/CuraEngineBackend/ProcessSlicedLayersJob.py
@@ -24,7 +24,7 @@ from cura import LayerPolygon
 
 import numpy
 from time import time
-from cura.UI.ExtrudersModel import ExtrudersModel
+from cura.Machines.Models.ExtrudersModel import ExtrudersModel
 catalog = i18nCatalog("cura")
