+++ plugins/UM3NetworkPrinting/src/Cloud/CloudOutputDevice.py
@@ -18,7 +18,7 @@ from UM.Scene.SceneNode import SceneNode
 
 from cura.CuraApplication import CuraApplication
 from cura.PrinterOutput.NetworkedPrinterOutputDevice import AuthState, NetworkedPrinterOutputDevice
-from cura.UI.PrinterOutputModel import PrinterOutputModel
+from cura.PrinterOutput.Models.PrinterOutputModel import PrinterOutputModel
 from cura.PrinterOutput.PrinterOutputDevice import ConnectionType
 
 from .CloudOutputController import CloudOutputController
