+++ plugins/UM3NetworkPrinting/src/Cloud/Models/CloudClusterPrinterStatus.py
@@ -3,7 +3,7 @@
 from typing import List, Union, Dict, Optional, Any
 
 from cura.PrinterOutput.PrinterOutputController import PrinterOutputController
-from cura.UI.PrinterOutputModel import PrinterOutputModel
+from cura.PrinterOutput.Models.PrinterOutputModel import PrinterOutputModel
 from .CloudClusterBuildPlate import CloudClusterBuildPlate
 from .CloudClusterPrintCoreConfiguration import CloudClusterPrintCoreConfiguration
 from .BaseCloudModel import BaseCloudModel
