+++ tests/PrinterOutput/Models/TestPrintJobOutputModel.py
@@ -1,35 +1,32 @@
-
-
 from unittest.mock import MagicMock
 
 import pytest
 
-from cura.UI.PrintJobOutputModel import PrintJobOutputModel
-from cura.UI.PrinterOutputModel import PrinterOutputModel
+from cura.PrinterOutput.Models.PrinterConfigurationModel import PrinterConfigurationModel
+from cura.PrinterOutput.Models.PrintJobOutputModel import PrintJobOutputModel
+from cura.PrinterOutput.Models.PrinterOutputModel import PrinterOutputModel
 
 test_validate_data_get_set = [
-    {"attribute": "name", "value": "YAY"},
-    {"attribute": "targetBedTemperature", "value": 192},
+    {"attribute": "compatibleMachineFamilies", "value": ["yay"]},
 ]
 
 test_validate_data_get_update = [
-    {"attribute": "isPreheating", "value": True},
-    {"attribute": "type", "value": "WHOO"},
-    {"attribute": "buildplate", "value": "NFHA"},
+    {"attribute": "configuration", "value": PrinterConfigurationModel()},
+    {"attribute": "owner", "value": "WHOO"},
+    {"attribute": "assignedPrinter", "value": PrinterOutputModel(MagicMock())},
     {"attribute": "key", "value": "YAY"},
     {"attribute": "name", "value": "Turtles"},
-    {"attribute": "bedTemperature", "value": 200},
-    {"attribute": "targetBedTemperature", "value": 9001},
-    {"attribute": "activePrintJob", "value": PrintJobOutputModel(MagicMock())},
-    {"attribute": "state", "value": "BEEPBOOP"},
+    {"attribute": "timeTotal", "value": 10},
+    {"attribute": "timeElapsed", "value": 20},
+    {"attribute": "state", "value": "BANANNA!"},
 ]
 
 
 @pytest.mark.parametrize("data", test_validate_data_get_set)
 def test_getAndSet(data):
-    model = PrinterOutputModel(MagicMock())
+    model = PrintJobOutputModel(MagicMock())
 
-    # Convert the first letter into a capital
+    # Convert the first letter into a capital   
     attribute = list(data["attribute"])
     attribute[0] = attribute[0].capitalize()
     attribute = "".join(attribute)
@@ -55,7 +52,7 @@ def test_getAndSet(data):
 
 @pytest.mark.parametrize("data", test_validate_data_get_update)
 def test_getAndUpdate(data):
-    model = PrinterOutputModel(MagicMock())
+    model = PrintJobOutputModel(MagicMock())
 
     # Convert the first letter into a capital
     attribute = list(data["attribute"])
