+++ tests/Settings/TestSettingVisibilityPresets.py
@@ -5,7 +5,7 @@ import os.path
 from UM.Preferences import Preferences
 from UM.Resources import Resources
 from cura.CuraApplication import CuraApplication
-from cura.UI.MachineModels.SettingVisibilityPresetsModel import SettingVisibilityPresetsModel
+from cura.Machines.Models.SettingVisibilityPresetsModel import SettingVisibilityPresetsModel
 from cura.Settings.SettingVisibilityPreset import SettingVisibilityPreset
 
 setting_visibility_preset_test_settings = {"test", "zomg", "derp", "yay", "whoo"}
